+++ src/main.cpp
@@ -36,7 +36,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.3.4"};
+constexpr auto version = std::string_view{"0.3.5"};
 
 struct BootstrapConfigResult final
 {